server/src/services/python/requirements_pandasai.txt
pandasai==2.0.42
fastapi==0.110.0
uvicorn==0.27.1
pandas==1.5.3
python-dotenv==1.0.1
supabase==2.3.5
httpx==0.24.1
gotrue==2.4.1
openai==1.12.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.3
numpy==1.26.4
pyyaml==6.0.2
redis==5.0.1